feat: 테스트 스니펫과 테스트 내용을 분
