+++ src/test/java/nextstep/subway/auth/acceptance/AuthAcceptanceTest.java
@@ -17,7 +17,7 @@ import static org.junit.jupiter.api.DynamicTest.dynamicTest;
 import static org.junit.jupiter.api.TestInstance.Lifecycle.PER_CLASS;
 
 @TestInstance(PER_CLASS)
-public class AuthAcceptanceTest extends AcceptanceTest {
+class AuthAcceptanceTest extends AcceptanceTest {
 
     private static final String EMAIL = "jordy-torvalds@jordy-torvalds.o-r.kr";
     private static final String CORRECT_PASSWORD = "jordy";
