+++ src/test/java/nextstep/subway/line/acceptance/LineSectionAcceptanceTest.java
@@ -1,29 +1,25 @@
 package nextstep.subway.line.acceptance;
 
-import io.restassured.RestAssured;
-import io.restassured.response.ExtractableResponse;
-import io.restassured.response.Response;
 import nextstep.subway.AcceptanceTest;
 import nextstep.subway.line.dto.LineRequest;
 import nextstep.subway.line.dto.LineResponse;
-import nextstep.subway.line.dto.SectionRequest;
-import nextstep.subway.station.StationAcceptanceTest;
 import nextstep.subway.station.dto.StationResponse;
-import org.junit.jupiter.api.*;
-import org.junit.jupiter.api.function.Executable;
-import org.springframework.http.HttpStatus;
-import org.springframework.http.MediaType;
+import org.junit.jupiter.api.BeforeEach;
+import org.junit.jupiter.api.DisplayName;
+import org.junit.jupiter.api.DynamicTest;
+import org.junit.jupiter.api.TestFactory;
 
 import java.util.List;
-import java.util.stream.Collectors;
 import java.util.stream.Stream;
 
 import static java.util.Arrays.asList;
-import static org.assertj.core.api.Assertions.assertThat;
+import static nextstep.subway.line.domain.LineSectorTestSnippet.*;
+import static nextstep.subway.line.domain.LineTestSnippet.지하철_노선_등록되어_있음;
+import static nextstep.subway.station.StationAcceptanceTest.지하철역_등록되어_있음;
 import static org.junit.jupiter.api.DynamicTest.dynamicTest;
 
 @DisplayName("지하철 구간 관련 기능")
-public class LineSectionAcceptanceTest extends AcceptanceTest {
+class LineSectionAcceptanceTest extends AcceptanceTest {
     private LineResponse 신분당선;
     private StationResponse 강남역;
     private StationResponse 양재역;
@@ -34,13 +30,13 @@ public class LineSectionAcceptanceTest extends AcceptanceTest {
     public void setUp() {
         super.setUp();
 
-        강남역 = StationAcceptanceTest.지하철역_등록되어_있음("강남역").as(StationResponse.class);
-        양재역 = StationAcceptanceTest.지하철역_등록되어_있음("양재역").as(StationResponse.class);
-        정자역 = StationAcceptanceTest.지하철역_등록되어_있음("정자역").as(StationResponse.class);
-        광교역 = StationAcceptanceTest.지하철역_등록되어_있음("광교역").as(StationResponse.class);
+        강남역 = 지하철역_등록되어_있음("강남역").as(StationResponse.class);
+        양재역 = 지하철역_등록되어_있음("양재역").as(StationResponse.class);
+        정자역 = 지하철역_등록되어_있음("정자역").as(StationResponse.class);
+        광교역 = 지하철역_등록되어_있음("광교역").as(StationResponse.class);
 
         LineRequest lineRequest = new LineRequest("신분당선", "bg-red-600", 강남역.getId(), 광교역.getId(), 10);
-        신분당선 = LineAcceptanceTest.지하철_노선_등록되어_있음(lineRequest).as(LineResponse.class);
+        신분당선 = 지하철_노선_등록되어_있음(lineRequest).as(LineResponse.class);
     }
 
     @DisplayName("지하철 노선을 관리한다")
@@ -103,7 +99,7 @@ public class LineSectionAcceptanceTest extends AcceptanceTest {
     Stream<DynamicTest> removeLineSection1() {
         // given
         List<StationResponse> 노선_내_정렬된_역 = asList(강남역, 정자역, 광교역);
-        
+
         return Stream.of(
                 dynamicTest("지하철 노선에 역 등록 요청 및 성공 확인(신분당선, 강남역 - 양재역)", 지하철_노선에_지하철역_등록_요청_및_성공_확인(신분당선, 강남역, 양재역, 2)),
                 dynamicTest("지하철 노선에 역 등록 요청 및 성공 확인(신분당선, 양재역 - 정자역)", 지하철_노선에_지하철역_등록_요청_및_성공_확인(신분당선, 양재역, 정자역, 2)),
@@ -117,102 +113,5 @@ public class LineSectionAcceptanceTest extends AcceptanceTest {
     @TestFactory
     void removeLineSection2() {
         dynamicTest("지하철_노선에_지하철역_제외_요청 및 실패 확인(신분당선, 양재역)", 지하철_노선에_지하철역_제외_요청_및_실패_확인(신분당선, 강남역));
-    }
-
-
-
-    public Executable 지하철_노선에_지하철역_등록_요청_및_성공_확인(LineResponse line, StationResponse upStation, StationResponse downStation, int distance) {
-        return () -> {
-            ExtractableResponse<Response> response =  지하철_노선에_지하철역_등록_요청(line, upStation, downStation, distance);
-
-            지하철_노선에_지하철역_등록됨(response);
-        };
-    }
-
-    private Executable 지하철_노선에_지하철역_등록_요청_및_실패_확인(LineResponse line, StationResponse upStation, StationResponse downStation, int distance) {
-        return () -> {
-            ExtractableResponse<Response> response =  지하철_노선에_지하철역_등록_요청(line, upStation, downStation, distance);
-
-            지하철_노선에_지하철역_등록_실패됨(response);
-        };
-    }
-
-    private Executable 지하철_노선_조회_요청_및_확인(LineResponse response, List<StationResponse> 노선_내_정렬된_역) {
-        return () -> {
-            ExtractableResponse<Response> getResponse = LineAcceptanceTest.지하철_노선_조회_요청(response);
-
-            지하철_노선_조회됨(getResponse);
-
-            지하철_노선에_지하철역_순서_정렬됨(getResponse, 노선_내_정렬된_역);
-        };
-    }
-
-    private Executable 지하철_노선에_지하철역_제외_요청_및_성공_확인(LineResponse line, StationResponse station) {
-        return () -> {
-            ExtractableResponse<Response> deleteResponse = this.지하철_노선에_지하철역_제외_요청(line, station);
-
-            지하철_노선에_지하철역_제외됨(deleteResponse);
-        };
-    }
-
-    private Executable 지하철_노선에_지하철역_제외_요청_및_실패_확인(LineResponse line, StationResponse station) {
-        return () -> {
-            ExtractableResponse<Response> deleteResponse = this.지하철_노선에_지하철역_제외_요청(line, station);
-
-            지하철_노선에_지하철역_제외_실패됨(deleteResponse);
-        };
-    }
-
-    public static ExtractableResponse<Response> 지하철_노선에_지하철역_등록_요청(LineResponse line, StationResponse upStation, StationResponse downStation, int distance) {
-        SectionRequest sectionRequest = new SectionRequest(upStation.getId(), downStation.getId(), distance);
-
-        return RestAssured
-                .given().log().all()
-                .contentType(MediaType.APPLICATION_JSON_VALUE)
-                .body(sectionRequest)
-                .when().post("/lines/{lineId}/sections", line.getId())
-                .then().log().all()
-                .extract();
-    }
-
-    private void 지하철_노선에_지하철역_등록됨(ExtractableResponse<Response> response) {
-        assertThat(response.statusCode()).isEqualTo(HttpStatus.OK.value());
-    }
-
-    private void 지하철_노선에_지하철역_등록_실패됨(ExtractableResponse<Response> response) {
-        assertThat(response.statusCode()).isEqualTo(HttpStatus.INTERNAL_SERVER_ERROR.value());
-    }
-    
-    private void 지하철_노선_조회됨(ExtractableResponse<Response> response) {
-        assertThat(response.statusCode()).isEqualTo(HttpStatus.OK.value());
-    }
-    
-    private void 지하철_노선에_지하철역_순서_정렬됨(ExtractableResponse<Response> response, List<StationResponse> expectedStations) {
-        LineResponse line = response.as(LineResponse.class);
-        List<Long> stationIds = line.getStations().stream()
-                .map(it -> it.getId())
-                .collect(Collectors.toList());
-
-        List<Long> expectedStationIds = expectedStations.stream()
-                .map(it -> it.getId())
-                .collect(Collectors.toList());
-
-        assertThat(stationIds).containsExactlyElementsOf(expectedStationIds);
-    }
-
-    private ExtractableResponse<Response> 지하철_노선에_지하철역_제외_요청(LineResponse line, StationResponse station) {
-        return RestAssured
-                .given().log().all()
-                .when().delete("/lines/{lineId}/sections?stationId={stationId}", line.getId(), station.getId())
-                .then().log().all()
-                .extract();
-    }
-
-    private void 지하철_노선에_지하철역_제외됨(ExtractableResponse<Response> response) {
-        assertThat(response.statusCode()).isEqualTo(HttpStatus.OK.value());
-    }
-
-    private void 지하철_노선에_지하철역_제외_실패됨(ExtractableResponse<Response> response) {
-        assertThat(response.statusCode()).isEqualTo(HttpStatus.INTERNAL_SERVER_ERROR.value());
     }
 }
